+++ src/language/markdown-source-code.js
@@ -13,7 +13,7 @@ import {
 	ConfigCommentParser,
 	Directive,
 } from "@eslint/plugin-kit";
-import { findOffsets } from "../util.js";
+import { lineEndingPattern } from "../util.js";
 
 //-----------------------------------------------------------------------------
 // Types
@@ -66,53 +66,38 @@ class InlineConfigComment {
 /**
  * Extracts inline configuration comments from an HTML node.
  * @param {Html} node The HTML node to extract comments from.
+ * @param {MarkdownSourceCode} sourceCode The Markdown source code object.
  * @returns {Array<InlineConfigComment>} The inline configuration comments found in the node.
  */
-function extractInlineConfigCommentsFromHTML(node) {
+function extractInlineConfigCommentsFromHTML(node, sourceCode) {
 	if (!configCommentStart.test(node.value)) {
 		return [];
 	}
+
+	/** @type {Array<InlineConfigComment>} */
 	const comments = [];
 
+	/** @type {RegExpExecArray} */
 	let match;
 
 	while ((match = htmlComment.exec(node.value))) {
 		if (configCommentStart.test(match[0])) {
-			const comment = match[0];
-
-			// calculate location of the comment inside the node
-			const start = {
-				...node.position.start,
-			};
-
-			const end = {
-				...node.position.start,
-			};
-
-			const {
-				lineOffset: startLineOffset,
-				columnOffset: startColumnOffset,
-			} = findOffsets(node.value, match.index);
-
-			start.line += startLineOffset;
-			start.column += startColumnOffset;
-			start.offset += match.index;
-
-			const commentLineCount = comment.split("\n").length - 1;
-
-			end.line = start.line + commentLineCount;
-			end.column =
-				commentLineCount === 0
-					? start.column + comment.length
-					: comment.length - comment.lastIndexOf("\n");
-			end.offset = start.offset + comment.length;
+			// calculate offset of the comment inside the node
+			const startOffset = match.index + node.position.start.offset;
+			const endOffset = startOffset + match[0].length;
 
 			comments.push(
 				new InlineConfigComment({
 					value: match[1].trim(),
 					position: {
-						start,
-						end,
+						start: {
+							...sourceCode.getLocFromIndex(startOffset),
+							offset: startOffset,
+						},
+						end: {
+							...sourceCode.getLocFromIndex(endOffset),
+							offset: endOffset,
+						},
 					},
 				}),
 			);
@@ -168,7 +153,7 @@ export class MarkdownSourceCode extends TextSourceCodeBase {
 	 * @param {Root} options.ast The root AST node.
 	 */
 	constructor({ text, ast }) {
-		super({ ast, text });
+		super({ ast, text, lineEndingPattern });
 		this.ast = ast;
 
 		// need to traverse the source code to get the inline config nodes
@@ -191,8 +176,8 @@ export class MarkdownSourceCode extends TextSourceCodeBase {
 	 */
 	getInlineConfigNodes() {
 		if (!this.#inlineConfigComments) {
-			this.#inlineConfigComments = this.#htmlNodes.flatMap(
-				extractInlineConfigCommentsFromHTML,
+			this.#inlineConfigComments = this.#htmlNodes.flatMap(htmlNode =>
+				extractInlineConfigCommentsFromHTML(htmlNode, this),
 			);
 		}