+++ tests/util.test.js
@@ -8,57 +8,13 @@
 //------------------------------------------------------------------------------
 
 import assert from "node:assert";
-import {
-	findOffsets,
-	frontmatterHasTitle,
-	stripHtmlComments,
-} from "../src/util.js";
+import { frontmatterHasTitle, stripHtmlComments } from "../src/util.js";
 
 //------------------------------------------------------------------------------
 // Tests
 //------------------------------------------------------------------------------
 
 describe("util", () => {
-	describe("findOffsets()", () => {
-		it("should return correct offsets for a simple string", () => {
-			const text = "Hello world!";
-			const offset = 6; // 'w' in "world"
-			const result = findOffsets(text, offset);
-			assert.deepStrictEqual(result, { lineOffset: 0, columnOffset: 6 });
-		});
-
-		it("should handle line breaks correctly", () => {
-			const text = "Hello\nworld!";
-			const offset = 6; // 'w' in "world"
-			const result = findOffsets(text, offset);
-			assert.deepStrictEqual(result, { lineOffset: 1, columnOffset: 0 });
-		});
-
-		it("should handle Windows-style line endings", () => {
-			const text = "Hello\r\nworld!";
-			const offset = 7; // 'w' in "world"
-			const result = findOffsets(text, offset);
-			assert.deepStrictEqual(result, { lineOffset: 1, columnOffset: 0 });
-		});
-
-		it("should handle offsets at the start of the string", () => {
-			const text = "Hello, world!";
-			const offset = 0; // Start of the string
-			const result = findOffsets(text, offset);
-			assert.deepStrictEqual(result, { lineOffset: 0, columnOffset: 0 });
-		});
-
-		it("should handle offsets at the end of the string", () => {
-			const text = "Hello, world!";
-			const offset = text.length - 1; // Last character '!'
-			const result = findOffsets(text, offset);
-			assert.deepStrictEqual(result, {
-				lineOffset: 0,
-				columnOffset: text.length - 1,
-			});
-		});
-	});
-
 	describe("frontmatterHasTitle()", () => {
 		const pattern = /^title:\s*My Document$/u;
 
@@ -80,7 +36,25 @@ describe("util", () => {
 			);
 		});
 
-		it("should return true if the pattern matches any line in multiline frontmatter", () => {
+		it("should return true if the pattern matches any line in multiline frontmatter (CRLF)", () => {
+			const frontmatter = [
+				"description: Test",
+				"title: My Document",
+				"author: lumirlumir",
+			].join("\r\n");
+			assert.strictEqual(frontmatterHasTitle(frontmatter, pattern), true);
+		});
+
+		it("should return true if the pattern matches any line in multiline frontmatter (CR)", () => {
+			const frontmatter = [
+				"description: Test",
+				"title: My Document",
+				"author: lumirlumir",
+			].join("\r");
+			assert.strictEqual(frontmatterHasTitle(frontmatter, pattern), true);
+		});
+
+		it("should return true if the pattern matches any line in multiline frontmatter (LF)", () => {
 			const frontmatter = [
 				"description: Test",
 				"title: My Document",